Version Beta 1.3.1

+++ cogs/autoverify.py
@@ -14,57 +14,69 @@ class Autoverify(commands.Cog):
     @commands.has_permissions(administrator=True)
     async def setmainrole(self, ctx, roleid):
         """Setze die Rolle die neue User bekommen"""
-        with open('./data/roles/mainrole.json', 'r') as f:
-            roles = json.load(f)
+        if ctx.guild.get_role(roleid) is not None:
+            with open('./data/roles/mainrole.json', 'r') as f:
+                roles = json.load(f)
 
-        roles[str(ctx.guild.id)] = int(roleid)
+            roles[str(ctx.guild.id)] = int(roleid)
 
-        with open('./data/roles/mainrole.json', 'w') as f:
-            json.dump(roles, f, indent=4)
-        await bp.delete_cmd(ctx)
-        await ctx.send("Main-Rolle gesetzt.", delete_after=bp.deltime)
+            with open('./data/roles/mainrole.json', 'w') as f:
+                json.dump(roles, f, indent=4)
+            await bp.delete_cmd(ctx)
+            await ctx.send("Main-Rolle gesetzt.", delete_after=bp.deltime)
+        else:
+            pass  # TODO ERROR
 
     @commands.command()
     @commands.has_permissions(administrator=True)
     async def setspacerone(self, ctx, roleid):
         """Setze die Spacerrolle 1, welche jeder User bekommen soll."""
-        with open('./data/roles/spacerone.json', 'r') as f:
-            roles = json.load(f)
+        if ctx.guild.get_role(roleid) is not None:
+            with open('./data/roles/spacerone.json', 'r') as f:
+                roles = json.load(f)
 
-        roles[str(ctx.guild.id)] = int(roleid)
+            roles[str(ctx.guild.id)] = int(roleid)
 
-        with open('./data/roles/spacerone.json', 'w') as f:
-            json.dump(roles, f, indent=4)
-        await bp.delete_cmd(ctx)
-        await ctx.send("Spacer 1 Rolle gesetzt.", delete_after=bp.deltime)
+            with open('./data/roles/spacerone.json', 'w') as f:
+                json.dump(roles, f, indent=4)
+            await bp.delete_cmd(ctx)
+            await ctx.send("Spacer 1 Rolle gesetzt.", delete_after=bp.deltime)
+        else:
+            pass  # TODO ERROR
 
     @commands.command()
     @commands.has_permissions(administrator=True)
     async def setspacertwo(self, ctx, roleid):
         """Setze die Spacerrolle 2, welche jeder User bekommen soll."""
-        with open('./data/roles/spacertwo.json', 'r') as f:
-            roles = json.load(f)
+        if ctx.guild.get_role(roleid) is not None:
+            with open('./data/roles/spacertwo.json', 'r') as f:
+                roles = json.load(f)
 
-        roles[str(ctx.guild.id)] = int(roleid)
+            roles[str(ctx.guild.id)] = int(roleid)
 
-        with open('./data/roles/spacertwo.json', 'w') as f:
-            json.dump(roles, f, indent=4)
-        await bp.delete_cmd(ctx)
-        await ctx.send("Spacer 2 Rolle gesetzt.", delete_after=bp.deltime)
+            with open('./data/roles/spacertwo.json', 'w') as f:
+                json.dump(roles, f, indent=4)
+            await bp.delete_cmd(ctx)
+            await ctx.send("Spacer 2 Rolle gesetzt.", delete_after=bp.deltime)
+        else:
+            pass  # TODO ERROR
 
     @commands.command()
     @commands.has_permissions(administrator=True)
     async def setspacerthree(self, ctx, roleid):
         """Setze die Spacerrolle 3, welche jeder User bekommen soll."""
-        with open('./data/roles/spacerthree.json', 'r') as f:
-            roles = json.load(f)
+        if ctx.guild.get_role(roleid) is not None:
+            with open('./data/roles/spacerthree.json', 'r') as f:
+                roles = json.load(f)
 
-        roles[str(ctx.guild.id)] = int(roleid)
+            roles[str(ctx.guild.id)] = int(roleid)
 
-        with open('./data/roles/spacerthree.json', 'w') as f:
-            json.dump(roles, f, indent=4)
-        await bp.delete_cmd(ctx)
-        await ctx.send("Spacer 3 Rolle gesetzt.", delete_after=bp.deltime)
+            with open('./data/roles/spacerthree.json', 'w') as f:
+                json.dump(roles, f, indent=4)
+            await bp.delete_cmd(ctx)
+            await ctx.send("Spacer 3 Rolle gesetzt.", delete_after=bp.deltime)
+        else:
+            pass  # TODO ERROR
 
     @commands.Cog.listener()
     async def on_message(self, message):
